requirements.txt
Django==4.1
djangorestframework==3.14.0
django-filter==22.1
drf-yasg==1.21.5
